+++ src/days/day06.rs
@@ -1,4 +1,3 @@
-use colored::Colorize;
 use std::collections::HashSet;
 use std::fs::File;
 use std::hash::Hash;
@@ -49,7 +48,7 @@ where
     Ok(io::BufReader::new(file).lines())
 }
 
-fn part1(path: &str) -> i32 {
+fn part1(path: &str) -> i64 {
     let start_chars: Vec<char> = vec!['^', '>', 'v', '<'];
     let mut matrix: Vec<Vec<char>> = Vec::<Vec<char>>::new();
     let mut start_pos: (usize, usize) = (0, 0);
@@ -77,7 +76,7 @@ fn part1(path: &str) -> i32 {
     return -1;
 }
 
-fn part2(path: &str) -> i32 {
+fn part2(path: &str) -> i64 {
     let start_chars: Vec<char> = vec!['^', '>', 'v', '<'];
     let mut matrix: Vec<Vec<char>> = Vec::<Vec<char>>::new();
     let mut start_pos: (usize, usize) = (0, 0);
@@ -121,7 +120,7 @@ fn part2(path: &str) -> i32 {
     return -1;
 }
 
-pub fn solve(part: i32) -> i32 {
+pub fn solve(part: i64) -> i64 {
     match part {
         1 => part1(INPUT_FILE_PATH),
         2 => part2(INPUT_FILE_PATH),
@@ -464,14 +463,14 @@ mod tests {
     #[test]
     fn example_1() {
         let result = part1(EXAMPLE_FILE_PATH);
-        let expected: i32 = 41;
+        let expected: i64 = 41;
         assert_eq!(result, expected);
     }
 
     #[test]
     fn example_2() {
         let result = part2(EXAMPLE_FILE_PATH);
-        let expected: i32 = 6;
+        let expected: i64 = 6;
         assert_eq!(result, expected);
     }
 }
